requirements.txt
cellseg==0.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
entrypoints==0.3
fonttools==4.38.0
grad-cam==1.5.4
idna==3.4
imageio==2.23.0
#install==1.3.5
joblib==1.3.1
kiwisolver==1.4.4
matplotlib~=3.10.3
networkx==2.8.8
numpy~=2.2.5
opencv-python~=4.11.0.86
packaging==22.0
pandas~=2.2.3
Pillow~=11.2.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.4.1
read-roi==1.6.0
requests==2.28.1
scikit-image==0.19.3
scikit-learn~=1.6.1
scipy~=1.15.3
six==1.16.0
#sklearn==0.0.post5
# Editable install with no version control (syotil==0.0.1)
#-e /home/yfong/deeplearning/syotil
testpath==0.5.0
threadpoolctl==3.1.0
tifffile==2022.10.10
torch~=2.7.0
torchvision~=0.22.0
tqdm==4.64.1
-e git+https://github.com/youyifong/tsp.git@e6dd315c902a03e7ad4879579f8e5dbd1cda344e#egg=tsp
ttach==0.0.3
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13

seaborn~=0.13.2